+++ zarrs_metadata_ext/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- **Breaking**: bump `zarrs_metadata` to 0.7.0
+
 ## [0.2.2] - 2025-10-26
 
 ### Fixed